pyproject.toml
[tool.poetry]
name = "team-challeng"
version = "0.1.0"
description = ""
authors = ["Grigoriy1209 <grigoriyv1209@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.7"



[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"
isort = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
